Move compatibility functions into different implementation files.
