+++ test/benchmark-common-tasks/Makefile
@@ -1,6 +1,6 @@
 check_PROGRAMS = bench-tasks
 
-bench_tasks_SOURCES = main.cpp
+bench_tasks_SOURCES = main.cpp compatibility.cpp
 bench_tasks_LIBS = ../util/test-util.a
 
 include ../../src/generic.mk
